Miscellaneous cleanup: renames, dead code removal, dedup
- Rename GetContainerLogs → GetContainerLogsFromDocker (DockerPlugin)
  and GetContainerLogsFromLoki (LokiPlugin) to disambiguate sources.
  MCP tool name preserved via [McpServerTool(Name = "GetContainerLogs")]
- Move PagedResult<T> from ConversationsController to Models/PagedResult.cs
- Remove dead MikroTikConfiguration fields (Host, ApiPort, Username,
  Password) — plugin uses Prometheus only
- Extract shared FormatKnowledgeFacts in KnowledgePlugin, eliminating
  duplicate formatting block between RecallKnowledge and SmartRecallKnowledge
