+++ src/HomelabBot/Controllers/TelemetryController.cs
@@ -1,4 +1,5 @@
 using HomelabBot.Data;
+using HomelabBot.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
