+++ src/HomelabBot/Models/PagedResult.cs
@@ -1,0 +1,9 @@
+namespace HomelabBot.Models;
+
+public record PagedResult<T>
+{
+    public required IReadOnlyList<T> Items { get; init; }
+    public int TotalCount { get; init; }
+    public int Page { get; init; }
+    public int PageSize { get; init; }
+}
